requirements.txt
# Index for CUDA 13.0 wheels (keeps PyPI as primary source)
--extra-index-url https://download.pytorch.org/whl/cu130

# Core dependencies
# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0

# Machine learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# Scientific computing
scipy>=1.10.0

# Configuration
pyyaml>=6.0

# CLI framework
click>=8.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# HTTP requests (for API calls)
requests>=2.28.0

# Testing
pytest>=7.0.0

# Progress bars
tqdm>=4.65.0

# Optional dependencies
# Time series forecasting
prophet>=1.1.0

# Hyperparameter optimization (optional)
# hyperopt>=0.2.7

# Deep learning (GPU, CUDA 13.0). For CPU-only, comment these three lines and install 'torch==2.9.1'.
torch==2.9.1+cu130
torchvision==0.24.1+cu130
torchaudio==2.9.1+cu130
